% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arp.R
\name{arp}
\alias{arp}
\title{Generate data with a piecewise constant mean and AR(p) noise.}
\usage{
arp(n, rho, mu = 0, sd = 1)
}
\arguments{
\item{n}{The number of data points.}

\item{rho}{A p-dimensional vector of correlation parameters (p >= 1).}

\item{mu}{The mean of the data -- this can be either a single value or a vector of length \code{n}. Defaults to 0.}

\item{sd}{The standard deviation of the noise (defaults to 1).}
}
\value{
A vector of length \code{n}.
}
\description{
Generate data with a piecewise constant mean and AR(p) noise.
}
\details{
Data is simulated from the model
\deqn{X_t = \mu_t + \rho (X_{t-1} - \mu_{t-1}) + \epsilon_t},
where \eqn{\epsilon_t \sim N(0, \sigma^2)}.
}
\examples{
set.seed(1)
n <- 100
rho <- 0.6
mu <- c(rep(0, 50), rep(1, 50))
x <- arp(n, rho, mu)
plot(x, type="l")

n <- 200
rho <- c(0.7, -0.2)
mu <- c(rep(0, 50), rep(1, 100), rep(2, 50))
x <- arp(n, rho, mu, 0.5)
plot(x, type="l")

}
