% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_segmentation.R
\name{binary_segmentation}
\alias{binary_segmentation}
\title{Estimate changepoints in mean, variance, or slope using binary segmentation}
\usage{
binary_segmentation(
  x,
  threshold = NULL,
  maxiter = NULL,
  model = "mean",
  loss = "lrs"
)
}
\arguments{
\item{x}{Vector of data.}

\item{threshold}{Threshold for determining changepoint candidates; defaults to \code{sqrt(2 * log(length(x)) * mad(x))}.}

\item{maxiter}{The maximum number of changepoints to find: defaults to the maximum (\code{length(x) - 2} if \code{model = "slope"}, otherwise \code{length(x) - 1}).}

\item{model}{Changepoint model: should be one of \code{"mean"}, \code{"var"}, or \code{"slope"}. The default is \code{"mean"}.}

\item{loss}{Loss function to use: for \code{model = "var"} only. One of \code{"cusum"}, \code{"icss"}, or \code{"lrs"} (default). See below for details.}
}
\value{
A list:
\itemize{
\item \code{results} Dataframe containing results.
\item \code{changepoints} Vector of changepoints detected.
\item \code{threshold} Value of \code{threshold} (either specified by user, or calculated in the function).
\item \code{maxiter} Value of \code{maxiter} (either specified by user, or default used by function).
}
}
\description{
Binary segmentation algorithm for detecting changes in mean, variance, or slope.
The algorithm terminates when either
\enumerate{
\item the maximum number of changepoints has been found; or
\item there are no remaining points where the CUSUM statistic is above the threshold.
}
}
\details{
There are three options for the changepoint model. In each case the data is assumed to be of the form
\eqn{x_t = \mu_t + \sigma_t \epsilon_t}, where \eqn{\epsilon_t \sim N(0, 1)}. The options are:
\itemize{
\item \code{model = "mean"} (change in mean): \eqn{\mu_t} is constant except at changepoints; \eqn{\sigma_t} is constant
\item \code{model = "slope"} (change in slope): \eqn{\mu_t = a_t + t b_t}, with \eqn{a_t} and \eqn{b_t} constant except at changepoints; \eqn{\sigma_t} is constant
\item \code{model = "var"} (change in variance): \eqn{\mu_t = 0} and \eqn{\sigma_t} is constant except at changepoints.
}
For the change in variance model (\code{model = "var"}), there are three possible loss functions for estimating changeoints.
If \code{loss = "lrs"}, the likelihood ratio statistic will be used for each interval
\deqn{(e - s + 1) \log \sum_{t=s}^e X_t^2 - (\tau - s + 1) \log \sum_{t=s}^{\tau} X_t^2 - (e - \tau) \log \sum_{t=\tau + 1}^e X_t^2}.
If \code{loss = "cusum"}, the CUSUM statistic of \eqn{X^2} will be used:
\deqn{(\frac{(\tau - s + 1)(e - \tau)}{e - s + 1})^2 (\frac{1}{\tau - s + 1} \sum_{t=s}^{\tau} X_t^2 + \frac{1}{e - \tau} \sum_{t=\tau + 1}^e X_t^2)}
If \code{loss = "icss"}, then
\deqn{\sum_{t=s}^{\tau} X_t^2 / \sum_{t=s}^e X_t^2 - \frac{\tau}{T}.}
}
\examples{
set.seed(10)
x <- c(rep(1,20), rep(-1,30), rep(1,50)) + rnorm(100)
binary_segmentation(x)

x <- c(seq(1, 10, length.out=50), rep(10, 50)) + rnorm(100)
binary_segmentation(x, model="slope")

x <- c(rnorm(50, sd=1), rnorm(50, sd=1.5), rnorm(50, sd=2))
binary_segmentation(x, model="var", loss="lrs", threshold=5)
binary_segmentation(x, model="var", loss="cusum", threshold=10)

}
