% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_interval.R
\name{calculate_interval}
\alias{calculate_interval}
\title{Calculate interval of \eqn{\phi} values which return a given set of changepoints.}
\usage{
calculate_interval(
  results,
  x = NULL,
  b0 = NULL,
  nu = NULL,
  nu2 = NULL,
  phi_obs = NULL,
  h1 = NULL,
  h2 = NULL,
  maxiter = NULL,
  C0 = NULL,
  C0_nu = NULL,
  autocor = FALSE,
  Sigma = NULL
)
}
\arguments{
\item{results}{Output of \code{find_changepoints}.}

\item{x}{Vector of data (if different from that supplied to \code{find_changepoints}).}

\item{b0}{Changepoint of interest; only needed if \code{model = "var"}, as otherwise this information is captured by the vector \code{nu}.
If this is not supplied, the first detected changepoint will be used.}

\item{nu}{Vector which defines the test statistic; not required if \code{model == "var"}. See below.}

\item{nu2}{Value of \eqn{||\nu||_2^2} (optional).}

\item{phi_obs}{Observed value of the test statistic (optional).}

\item{h1}{Left window size; only needed if \code{model = "var"}.}

\item{h2}{Right window size; only needed if \code{model = "var"}.}

\item{maxiter}{Maximum number of changepoints to detect.}

\item{C0}{Vector containing cumulative sums of \code{x} (if \code{model = "mean"} or code{model = "slope"}) or \code{x^2} (if \code{model = "var"}).}

\item{C0_nu}{Vector containing cumulative sums of \code{nu}.}

\item{autocor}{Logical: \code{TRUE} if data is autocorrelated, otherwise \code{FALSE} (default).}

\item{Sigma}{Covariance matrix of the data (only needed if \code{autocor = TRUE}).}
}
\value{
A 2-dimensional vector.
}
\description{
For a given changepoint algorithm, model, and dataset, calculate the interval of \eqn{\phi} values such that applying the algorithm to \eqn{X'(\phi)} will give
a particular set of changepoint estimates and directions of change. (Wrapper function for \code{calculate_interval_bs}, \code{calculate_interval_wbs}, and \code{calculate_interval_not}.)
}
\details{
For the change in mean (\code{model = "mean"}) and change in slope (\code{model = "slope"}), \eqn{X'(\phi)} is defined as
\deqn{X'(\phi) = X_{obs} + \frac{1}{||\nu||^2} \nu (\phi - \phi_{obs}).}
The vector \code{nu} must be supplied to the function.
For the change in variance model, \eqn{X'(\phi)} is defined by
\deqn{
X'(\phi) = \begin{cases}
X_t & t \leq \hat{\tau} - h_1 \text{ or } t \geq \hat{\tau} + h_2 + 1 \\
\sqrt{\frac{\phi}{\phi_{obs}}} X_t & \hat{\tau} - h_1 + 1 \leq t \leq \hat{\tau} \\
\sqrt{\frac{1 - \phi}{1 - \phi_{obs}}} X_t & \hat{\tau} + 1 \leq t \leq \hat{\tau} + h_2.
\end{cases}
}
The window sizes \code{h1} and \code{h2} must be supplied to the function.

The parameters \code{C0}, \code{nu2}, and \code{C0_nu} are optional, but will speed up computation time if the function is used multiple times (e.g. within \code{calculate_pvals_all}).
The latter two are only relevant if \code{model = "mean"} or \code{model = "slope"}.
}
\examples{
set.seed(100)
x <- rnorm(100) + c(rep(1,50), rep(-1,50))
results <- find_changepoints(x, "bs", params=list(threshold=4))
b <- results$changepoints
h <- 10
nu <- c(rep(0, b[1] - h), rep(1/h, h), rep(-1/h, h), rep(0, length(x) - b[1] - h))
calculate_interval(results, nu=nu)

}
