% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvals_all.R
\name{calculate_pvals_all}
\alias{calculate_pvals_all}
\title{Calculate p-values for changepoints}
\usage{
calculate_pvals_all(
  results,
  h = NULL,
  N = 1,
  num_pvals = NULL,
  sigma2 = NULL,
  include_original = TRUE,
  cp_bound = TRUE,
  gamma = 1,
  return_probs = FALSE
)
}
\arguments{
\item{results}{Output of \code{find_changepoints}.}

\item{h}{Window size: the null hypothesis is that there are no changepoints within a window \code{h} of the changepoint of interest. This can be a positive integer (>=2)
or a vector of length 2 (if the window is asymmetric) or \code{NULL} (if the window is defined by the changepoints either side of the changepoint of interest).}

\item{N}{Number of random \eqn{\psi}'s to generate; defaults to 1.}

\item{num_pvals}{Maximum number of p-values to calculate; if set to NA, it will default to \code{length(x) - 1}. P-values are calculated for changepoints
in the order of detection for iterative algorithms, otherwise they are calculated in the numeric order of changepoint locations.}

\item{sigma2}{Variance of \code{x}, for piecewise constant/linear mean models. If not given, it will be estimated using median absolute deviation.}

\item{include_original}{Whether to include the \eqn{\psi} value corresponding to the observed data in place of one of the random samples. Defaults to \code{TRUE}.}

\item{cp_bound}{If \code{TRUE}, then if there is an estimated changepoint within the window that is fixed under the null hypothesis,
this changepoint will be used as the boundary of the window.}

\item{gamma}{Value in \code{(0, 1]}. If \code{h = NULL}, then \eqn{h} is taken to be \code{gamma} times the distance between the changepoint
of interest and the estimated changepoints on either side. Ignored if \code{h} specified; defaults to \code{1}.}

\item{return_probs}{Whether to return the values of \eqn{Pr(\phi \in S \& |\phi| > |\phi_{obs}|)} and \eqn{Pr(\phi \in S)}
for each \eqn{\psi} as part of the output. Defaults to \code{FALSE}.}
}
\value{
A list.
\itemize{
\item \code{p_value} A vector of p-values
\item \code{P_both} If \code{return_probs=TRUE}, a matrix containing values of \eqn{Pr(\phi \in S \& |\phi| > |\phi_{obs}|)} for each
\eqn{\psi}; otherwise \code{NA}.
\item \code{P_phi_in_S} If \code{return_probs=TRUE}, a matrix containing values of \eqn{Pr(\phi \in S)} for each \eqn{\psi};
otherwise \code{NA}.
}
}
\description{
Calculate p-values for detected changepoints, for change in mean, slope, or variance. Changepoints can be detected using
the CUSUM statistic with either binary segmentation, wild binary segmentation, or narrowest over threshold.
}
\details{
Current options for changepoint models are change in mean (\code{model = "mean"}), change in slope (\code{model = "slope"}), and change in variance
(\code{model = "var"}). Either binary segmentation (\code{"bs"}), wild binary segmentation (\code{"wbs"}) or narrowest-over-threshold (\code{"not"}) can be used to estimate changepoints.
For the change in variance model, the CUSUM statistic must be used as the loss function (\code{loss = "cusum"}): otherwise, the p-values cannot be calculated exactly.
In this case p-values should be estimated using the function \code{calculate_pvals_var}.

Given a changepoint of interest \eqn{\tau_j}, there are several options for the null hypothesis:
\itemize{
\item There are no changepoints within a window size \eqn{h} of \eqn{\tau_j}. In this case \code{h} should be supplied. It is possible
to use different values of \code{h} on either side of \eqn{\tau_j}, in which case \code{h} should be a vector of length 2. If \eqn{\tau_j - h < 1},
then the value of \code{h} to the left of \eqn{\tau_j} will be set to \eqn{\tau_j}; similarly, if \eqn{\tau_h + h > n}, then the value of \code{h}
to the right of \eqn{\tau_j} will be set to \eqn{n - \tau_j}.
If \code{cp_bound = TRUE}, then if there is another estimated changepoint within \code{h} of \eqn{\tau_j} then the window size used
will be the minimum of \code{h} and the distance between \eqn{\tau_j} and the closest estimated changepoint to \eqn{\tau_j}.
\item There are no other changepoints between \eqn{\tau_{j-1}} and \eqn{\tau_{j+1}}. In this case \code{h} should be set to \code{NULL}
and \code{gamma} should be 1. (This is the default for the function.)
\item There are no other changepoints between the midpoint of \eqn{\tau_{j-1}} and \eqn{\tau_j} and the midpoint of \eqn{\tau_j} and
\eqn{\tau_{j+1}}. In this case use \code{h = NULL} and \code{gamma = 0.5}.
}
}
\examples{
set.seed(100)
x <- c(rep(1,50), rep(-1,50), rep(1,50), rep(-1,50)) + rnorm(200)
results <- find_changepoints(x, method="bs", params=list(threshold=4))
calculate_pvals_all(results, h=20, return_probs=TRUE)
calculate_pvals_all(results, h=20, N=5, return_probs=TRUE)

x <- c(rep(0, 100), seq(0, 10, length.out=100)) + rnorm(200)
results <- find_changepoints(x, method="bs", model="slope")
calculate_pvals_all(results, h=20, sigma2=1, return_probs=TRUE)
calculate_pvals_all(results, h=20, N=5, sigma2=1, return_probs=TRUE)

x <- c(rnorm(100), rnorm(100, sd=2))
results <- find_changepoints(x, method="bs", params=list(threshold=10, loss="cusum"), model="var")
calculate_pvals_all(results, h=20, return_probs=TRUE)
calculate_pvals_all(results, h=20, N=5, return_probs=TRUE)

}
