% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pvals_autocor.R
\name{calculate_pvals_autocor}
\alias{calculate_pvals_autocor}
\title{Calculate p-values for change in mean model, with autocorrelation}
\usage{
calculate_pvals_autocor(
  results,
  Sigma = NULL,
  rho = NULL,
  sigma2 = NULL,
  h = NULL,
  N = 1,
  est_sigma_pars = list(K = 10, rho_grid_size = 0.01),
  cp_bound = TRUE,
  gamma = 1,
  num_pvals = NULL
)
}
\arguments{
\item{results}{Output of \code{find_changepoints} applied to data.}

\item{Sigma}{Covariance matrix of the data. This can be calculated within the function (exactly if \code{rho} and \code{sigma2} are supplied, or estimated if they are not)
but if the function is run multiple times where the value of \code{Sigma} is the same, it will run faster if \code{Sigma} is calculated in advance.}

\item{rho}{Autocorrelation parameter for AR(1) model; ignored if \code{Sigma} is supplied.}

\item{sigma2}{Noise variance for AR(1) model; ignored if \code{Sigma} is supplied.}

\item{h}{Window size: either a positive integer (>=2) or vector of length 2 or \code{NULL}. If \code{NULL} then the changepoints either side of
the changepoint of interest are used to define the window.}

\item{N}{Number of random \eqn{\psi}'s to generate; defaults to 1.}

\item{est_sigma_pars}{List of parameters to be supplied to the function \code{est_rho}, in order to estimate the variance and autocorrelation parameter.
This will be ignored if either \code{Sigma} or both \code{rho} and \code{sigma2} are supplied.}

\item{cp_bound}{If \code{TRUE}, then if there is an estimated changepoint within the window that is fixed under the null hypothesis,
this changepoint will be used as the boundary of the window.}

\item{gamma}{Number in \code{(0, 1]}. If \code{h = NULL}, then \eqn{h} is taken to be \code{gamma} times the distance between the changepoint
of interest and the estimated changepoints on either side. Ignored if \code{h} specified; defaults to \code{1}.}

\item{num_pvals}{Number of p-values to compute. If \code{NULL}, the function will calculate p-values for all detected changepoints.}
}
\value{
A list:
\itemize{
\item \code{p_value} A vector containing p-values for each estimated changepoint.
\item \code{P_both} A matrix containing values of \eqn{Pr(\phi \in S \& |\phi| > |\phi_{obs}|)} for each estimated changepoint and sampled \eqn{\psi} value.
\item \code{P_phi_in_S} A matrix containing values of \eqn{Pr(\phi \in S)} for each estimated changepoint and sampled \eqn{\psi} value.
}
}
\description{
Calculate p-values for detected changepoints, for change in mean or slope, where there is autocorrelated noise. Changepoints can be detected using
the CUSUM statistic with either binary segmentation, wild binary segmentation, or narrowest over threshold.
}
\details{
The model is \eqn{X_t = \mu_t + Z_t}, where \eqn{\mu_t} is constant except at changepoints and \eqn{Z_t \sim N(0, \Sigma)}.
Either the covariance matrix \eqn{\Sigma} can be supplied to the function or, if \eqn{Z_t} follows an AR(1) model of the form
\eqn{Z_t = \rho Z_{t-1} + \sigma \epsilon_t} with eqn{\epsilon_t \sim N(0, 1)}, then \eqn{\rho} and \eqn{\sigma^2} can be supplied instead.
If neither of \code{Sigma} or \code{rho} is specified, then the function will estimate \eqn{rho} and \eqn{\sigma^2} under the assumption
that the data follows an AR(1) model.

Given a changepoint of interest \eqn{\tau_j}, there are several options for the null hypothesis:
\itemize{
\item There are no changepoints within a window size \eqn{h} of \eqn{\tau_j}. In this case \code{h} should be supplied. It is possible
to use different values of \code{h} on either side of \eqn{\tau_j}, in which case \code{h} should be a vector of length 2. If \eqn{\tau_j - h < 1},
then the value of \code{h} to the left of \eqn{\tau_j} will be set to \eqn{\tau_j}; similarly, if \eqn{\tau_h + h > n}, then the value of \code{h}
to the right of \eqn{\tau_j} will be set to \eqn{n - \tau_j}.
If \code{cp_bound = TRUE}, then if there is another estimated changepoint within \code{h} of \eqn{\tau_j} then the window size used
will be the minimum of \code{h} and the distance between \eqn{\tau_j} and the closest estimated changepoint to \eqn{\tau_j}.
\item There are no other changepoints between \eqn{\tau_{j-1}} and \eqn{\tau_{j+1}}. In this case \code{h} should be set to \code{NULL}
and \code{gamma} should be 1. (This is the default for the function.)
\item There are no other changepoints between the midpoint of \eqn{\tau_{j-1}} and \eqn{\tau_j} and the midpoint of \eqn{\tau_j} and
\eqn{\tau_{j+1}}. In this case use \code{h = NULL} and \code{gamma = 0.5}.
}
}
\examples{
set.seed(1)
x <- arp(100, 0.6, c(rep(0, 50), rep(2, 50)))
results <- find_changepoints(x, method="bs", model="mean")
calculate_pvals_autocor(results, rho=0.6, sigma2=1, h=20)
calculate_pvals_autocor(results, h=20) # using estimated variances

}
