% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_S.R
\name{calculate_S}
\alias{calculate_S}
\title{Calculate S}
\usage{
calculate_S(
  x,
  results,
  tau = NULL,
  nu = NULL,
  nu2 = NULL,
  h = NULL,
  phi_obs = NULL,
  first_cp_only = FALSE
)
}
\arguments{
\item{x}{Vector of data.}

\item{results}{Output of \code{find_changepoints}.}

\item{tau}{Changepoint of interest. Defaults to the first detected changepoint, if not given.}

\item{nu}{Vector that defines the test statistic, for \code{model = "mean"} or \code{model = "slope"}. See below.}

\item{nu2}{Value of \eqn{||\nu||_2^2}; optional.}

\item{h}{Window size; only needed if \code{model = "var"}. If the window is asymmetric, this should be a vector of length 2 containing the left and right window sizes.
Otherwise, it should be a single integer value.}

\item{phi_obs}{Observed value of the test statistic; optional.}

\item{first_cp_only}{Logical. If \code{TRUE}, condition on the fact that the changepoint of interest is in the model;
if \code{FALSE}, condition on all changepoints. Defaults to \code{FALSE}.}
}
\value{
A dataframe containing intervals with the changepoints obtained when \eqn{\phi} is in each interval.
}
\description{
For a given dataset and changepoint algorithm (one of binary segmentation, wild binary segmentation, or narrowest over threshold),
calculate the possible outcomes that can be obtained by applying the algorithm to
the perturbed dataset \eqn{X'(\phi)}, and the intervals of \eqn{\phi} values that lead to each possible outcome. See below for more details.
}
\details{
There are three options for the changepoint model:
\itemize{
\item Change in mean (\code{model = "mean"}): \eqn{X_t = \mu_t + \epsilon_t} where \eqn{\mu_t} is piecewise constant and \eqn{\epsilon_t \sim N(0, \sigma^2)}.
\item Change in slope (\code{model = "slope"}): same as above but with \eqn{\mu_t} piecewise linear.
\item Change in variance (\code{model = "var"}): \eqn{\mu_t} is assumed to be 0 for all \eqn{t}, and \eqn{\epsilon_t \sim N(0, \sigma_t^2)} where \eqn{\sigma_t}
is piecewise constant.
}

For the change in mean (\code{model = "mean"}) and change in slope (\code{model = "slope"}), \eqn{X'(\phi)} is defined as
\deqn{X'(\phi) = X_{obs} + \frac{1}{||\nu||^2} \nu (\phi - \phi_{obs}).}
The vector \code{nu} must be supplied to the function.
For the change in variance model, \eqn{X'(\phi)} is defined by
\deqn{
X'(\phi) = \begin{cases}
X_t & t \leq \hat{\tau} - h_1 \text{ or } t \geq \hat{\tau} + h_2 + 1 \\
\sqrt{\frac{\phi}{\phi_{obs}}} X_t & \hat{\tau} - h_1 + 1 \leq t \leq \hat{\tau} \\
\sqrt{\frac{1 - \phi}{1 - \phi_{obs}}} X_t & \hat{\tau} + 1 \leq t \leq \hat{\tau} + h_2.
\end{cases}
}
}
\examples{
set.seed(100)
x <- rnorm(100) + c(rep(1,50), rep(-1,50))
results <- find_changepoints(x, "bs")
b <- results$changepoints
print(b)
h <- 10
nu <- c(rep(0, b[1] - h), rep(1/h, h), rep(-1/h, h), rep(0, length(x) - b[1] - h))
calculate_S(x, results, nu=nu, first_cp_only=TRUE)


}
