% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_changepoints.R
\name{find_changepoints}
\alias{find_changepoints}
\title{Estimate changepoints in mean, variance, or slope.}
\usage{
find_changepoints(x, method, params = list(), model = "mean")
}
\arguments{
\item{x}{Vector of data.}

\item{method}{Changepoint method: one of \code{"bs"} (binary segmentation), \code{"wbs"} (wild binary segmentation), \code{"not"} (narrowest over threshold),
or \code{"pelt"} (PELT). For PELT the \code{changepoint} package must be installed.}

\item{params}{A list containing parameters for the changepoint algorithm (other than \code{x} and \code{model}). (See below.)
To use default values for all parameters, enter an empty list: \code{list()}.}

\item{model}{Changepoint model: one of \code{"mean"} (piecewise constant mean, default), \code{"slope"} (piecewise linear mean), or \code{"var"} (piecewise constant variance).}
}
\value{
A list:
\itemize{
\item \code{x} The data.
\item \code{results} Dataframe containing results.
\item \code{results_full} For wild binary segmentation or narrowest over threshold, a dataframe containing the changepoint candidates for every random interval.
\item \code{changepoints} Vector of changepoints detected.
\item \code{params} The parameters supplied to the function.
}
}
\description{
Function to estimate changes in mean, variance or slope for univariate data. Wrapper function for \code{binary_segmentation}, \code{wild_binary_segmentation},
and \code{narrowest_over_threshold}. Also includes PELT using the \code{changepoint} package.
}
\details{
The parameter \code{params} should be a list containing relevant parameters for the changepoint algorithm. This can be left empty, in which case parameters will take their
default values. There is no need to specify \code{x} or \code{model} as these are supplied directly to the \code{find_changepoints} function. See documentation for individual
functions for a list of possible parameters.
}
\examples{
set.seed(100)
x <- rnorm(100) + c(rep(1,45), rep(-1,10), rep(1,45))
results_bs <- find_changepoints(x, "bs", list(threshold=3))
print(results_bs)

results_wbs <- find_changepoints(x, "wbs", list(threshold=4, num_rand_samples=100))
print(results_wbs)

results_not <- find_changepoints(x, "not", list(threshold=4, num_rand_samples=100))
print(results_not)

results_pelt <- find_changepoints(x, "pelt", list())
print(results_pelt)

x <- c(rnorm(100), rnorm(100, sd=1.5))
find_changepoints(x, "bs", list(threshold=4), "var")

}
