% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad.R
\name{mad}
\alias{mad}
\title{Estimate variance of a vector by median absolute deviation, assuming piecewise linear mean}
\usage{
mad(x, lag = 1)
}
\arguments{
\item{x}{Vector of data.}

\item{lag}{Lag at which to calculate differences; defaults to 1.}
}
\value{
Estimated variance of \code{x}.
}
\description{
Estimate variance of a vector using median absolute deviation, under the assumption that the data
is generated with a piecewise constant mean plus Gaussian noise.
}
\examples{
x <- rnorm(200, sd=2) + c(rep(1, 100), rep(-1, 100))
mad(x)

}
