% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mad.R
\name{mad_slope}
\alias{mad_slope}
\title{Estimate variance of a vector by median absolute deviation, assuming piecewise constant mean}
\usage{
mad_slope(x, lag = 1)
}
\arguments{
\item{x}{Vector of data.}

\item{lag}{Lag at which to calculate differences; defaults to 1.}
}
\value{
Estimated variance of \code{x}.
}
\description{
Estimate variance of a vector using median absolute deviation, under the assumption that the data
is generated with a piecewise linear mean plus Gaussian noise.
}
\examples{
x <- rnorm(200, sd=2) + seq(0, 3, length.out=200)
mad_slope(x)

}
