% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/narrowest_over_threshold.R
\name{narrowest_over_threshold}
\alias{narrowest_over_threshold}
\title{Narrowest over threshold changepoint algorithm.}
\usage{
narrowest_over_threshold(
  x,
  num_rand_samples = 1000,
  random_samples = NULL,
  threshold = NULL,
  maxiter = NULL,
  model = "mean",
  loss = "lrs",
  return_full_results = FALSE
)
}
\arguments{
\item{x}{Vector of data.}

\item{num_rand_samples}{Number of random intervals to use (defaults to 1000). Ignored if \code{random_samples} is specified.}

\item{random_samples}{N x 2 matrix containing random intervals (optional; can be generated within the function).}

\item{threshold}{Changepoint detection threshold, should be a non-negative number. Defaults to \code{sqrt(2 * log(length(x)) * mad(x))}.}

\item{maxiter}{Maximum number of changepoints to return: defaults to \code{n - 1} (or \code{n - 2} if \code{model = "slope"}.}

\item{model}{Changepoint model: one of \code{"mean"} (change in mean), \code{"slope"} (change in slope), or \code{"var"} (change in variance).}

\item{loss}{Loss function (only if \code{model = "var"}): one of \code{"lrs"}, \code{"cusum"}, or \code{"icss"}. (See below.)}

\item{return_full_results}{If \code{TRUE}, a data frame containing the changepoint candidate for each random interval will be returned with the results.}
}
\value{
A list.
\itemize{
\item \code{results} Dataframe containing results
\item \code{changepoints} Vector of changepoints detected
\item \code{random_samples} N x 2 matrix containing random intervals used in the algorithm.
\item \code{threshold} Value of \code{threshold}
\item \code{maxiter} Value of \code{maxiter}
}
}
\description{
Apply the narrowest over threshold changepoint algorithm to a vector of data. Can be used to detect changes in mean, slope, or variance.
}
\details{
There are three options for the changepoint model. In each case the data is assumed to be of the form
\eqn{x_t = \mu_t + \sigma_t \epsilon_t}, where \eqn{\epsilon_t \sim N(0, 1)}. The options are:
\itemize{
\item \code{model = "mean"} (change in mean): \eqn{\mu_t} is constant except at changepoints; \eqn{\sigma_t} is constant
\item \code{model = "slope"} (change in slope): \eqn{\mu_t = a_t + t b_t}, with \eqn{a_t} and \eqn{b_t} constant except at changepoints; \eqn{\sigma_t} is constant
\item \code{model = "var"} (change in variance): \eqn{\mu_t = 0} and \eqn{\sigma_t} is constant except at changepoints.
}
For the change in variance model (\code{model = "var"}), there are three possible loss functions for estimating changeoints.
If \code{loss = "lrs"}, the likelihood ratio statistic will be used for each interval
\deqn{(e - s + 1) \log \sum_{t=s}^e X_t^2 - (\tau - s + 1) \log \sum_{t=s}^{\tau} X_t^2 - (e - \tau) \log \sum_{t=\tau + 1}^e X_t^2}.
If \code{loss = "cusum"}, the CUSUM statistic of \eqn{X^2} will be used:
\deqn{(\frac{(\tau - s + 1)(e - \tau)}{e - s + 1})^2 (\frac{1}{\tau - s + 1} \sum_{t=s}^{\tau} X_t^2 + \frac{1}{e - \tau} \sum_{t=\tau + 1}^e X_t^2)}
If \code{loss = "icss"}, then
\deqn{\sum_{t=s}^{\tau} X_t^2 / \sum_{t=s}^e X_t^2 - \frac{\tau}{T}.}
}
\examples{
set.seed(1)
x <- rnorm(200) + c(rep(1,90), rep(-1,20), rep(1,90))
results <- narrowest_over_threshold(x, num_rand_samples=500, threshold=4)
print(results$results)
print(results$changepoints)

}
