% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_changepoints.R
\name{plot_changepoints}
\alias{plot_changepoints}
\title{Plot data with estimated changepoint locations and estimated trends.}
\usage{
plot_changepoints(results, return_mu = FALSE, use_ggplot = FALSE, tvals = NULL)
}
\arguments{
\item{results}{Output of \code{find_changepoints} applied to a vector of data.}

\item{return_mu}{Whether to return estimates of the trend; defaults to \code{FALSE}.}

\item{use_ggplot}{If \code{TRUE}, creates a plot using \code{ggplot2} package; if \code{FALSE},
the plot will be created using R's \code{plot} function.}

\item{tvals}{Vector of values to plot on the x-axis.}
}
\value{
If \code{return_mu = TRUE}, vector of estimated parameters (mean, slope, or variance) for each changepoint segment.
}
\description{
Plot data with estimated changepoint locations and estimated trends.
}
\examples{
x <- c(rep(0, 100), rep(1, 100), rep(-1, 100), rep(1, 100), rep(2, 100)) + rnorm(500)
results <- find_changepoints(x, "bs")
plot_changepoints(results)

}
